Update skill mutations to add verification with GitHub

+++ packages/server/src/graphql/nexus/mutations/updateDesiredSkills.ts
@@ -1,5 +1,6 @@
 import { PromiseUtils } from "@makepurple/utils";
 import { arg, mutationField, nonNull } from "nexus";
+import { octokit } from "../../../services";
 
 export const updateDesiredSkills = mutationField("updateDesiredSkills", {
 	type: nonNull("UpdateDesiredSkillsPayload"),
@@ -9,7 +10,7 @@ export const updateDesiredSkills = mutationField("updateDesiredSkills", {
 	authorize: (parent, args, { user }) => {
 		return !!user;
 	},
-	resolve: async (parent, args, { prisma, user }) => {
+	resolve: async (parent, args, { octokit: graphql, prisma, user }) => {
 		if (!user) throw new Error();
 
 		const skillIds = args.data.skills
@@ -20,31 +21,70 @@ export const updateDesiredSkills = mutationField("updateDesiredSkills", {
 			.filter((skill) => !!skill.name_owner)
 			.map((skill) => skill.name_owner) as { name: string; owner: string }[];
 
-		const skillsById = await prisma.skill.findMany({
-			where: { id: { in: skillIds } }
+		const existingSkills = await prisma.skill.findMany({
+			where: {
+				OR: [{ id: { in: skillIds } }, ...skillNameOwners]
+			}
 		});
 
-		const updatedUser = await prisma.$transaction(async (transaction) => {
+		const toCreateSkills = skillNameOwners.filter(
+			({ name, owner }) =>
+				!existingSkills.some((skill) => skill.name === name && skill.owner === owner)
+		);
+
+		const verifySkill = async (name: string, owner: string): Promise<boolean> => {
+			const verified = await graphql`
+				query VerifySkill($name: String!, $owner: String!) {
+					repository(followRenames: false, name: $name, owner: $owner) {
+						id
+					}
+				}
+			`
+				.cast<octokit.VerifySkillQuery, octokit.VerifySkillQueryVariables>({
+					name,
+					owner
+				})
+				.then((result) => !!result.repository)
+				.catch(() => false);
+
+			return verified;
+		};
+
+		const verified = await PromiseUtils.some(
+			toCreateSkills,
+			{ concurrency: 2 },
+			async (skill) => await verifySkill(skill.name, skill.owner)
+		);
+
+		if (!verified) throw new Error("All skills must be from GitHub");
+
+		const record = await prisma.$transaction(async (transaction) => {
 			await transaction.user.update({
 				where: { id: user.id },
 				data: { desiredSkills: { deleteMany: {} } }
 			});
 
-			const skillsByNameOwner = await PromiseUtils.map(
-				skillNameOwners,
+			const newSkills = await PromiseUtils.map(
+				toCreateSkills,
 				{ concurrency: 2 },
 				async ({ name, owner }) => {
-					return await transaction.skill.upsert({
-						where: { name_owner: { name, owner } },
-						create: { name, owner },
-						update: {}
+					return await transaction.skill.create({
+						data: {
+							name,
+							organization: {
+								connectOrCreate: {
+									where: { name: owner },
+									create: { name: owner }
+								}
+							}
+						}
 					});
 				}
 			);
 
-			const skillsToConnect = [...skillsById, ...skillsByNameOwner];
+			const skillsToConnect = [...existingSkills, ...newSkills];
 
-			const record = await transaction.user.update({
+			return await transaction.user.update({
 				where: { id: user.id },
 				data: {
 					desiredSkills: {
@@ -63,10 +103,8 @@ export const updateDesiredSkills = mutationField("updateDesiredSkills", {
 					}
 				}
 			});
-
-			return { record };
 		});
 
-		return updatedUser;
+		return { record };
 	}
 });
